Rework triggering of vectorization (#15)
Remove compiled vectorization of binary search trees for starting characters. This was causing long compilation times for patterns that start with large character classes, e.g. "\w".

Add a RegexOptions.Vectorize option that will turn on vectorization. This is behind an option because support for vectorization in the System.Numerics.Vectors library is dependent on the runtime (needs 64-bit RyuJIT in .NET Core or Framework >=4.6).

Also add #if UNSAFE directives to make the project compile with unsafe turned off.